edge/requirements.txt
pandas>=2.0,<3.0
paho-mqtt>=1.6,<2.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
